# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: BuildZ API
  version: 0.2.1
  description: M0-M6 hackathon backend. Private authenticated workflows and explicit anonymous public projections. All
    times are UTC ISO-8601. Monetary amounts are integer minor units. Notifications use authenticated polling; payments
    are school-managed, not charged by BuildZ. Private file uploads are synchronous, at most 2 MiB per file/50 MiB per
    project, with signature checks but no malware-scan guarantee.
servers:
  - url: http://localhost:3001
    description: Local backend
security:
  - bearerAuth: []
paths:
...
  /v1/meta:
    get:
      operationId: getMeta
      summary: getMeta
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/ServiceMeta"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: []
  /v1/me:
    get:
      operationId: getMe
      summary: getMe
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Me"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMe
      summary: updateMe
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Me"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMe"
  /v1/institutions:
    get:
      operationId: listInstitutions
      summary: listInstitutions
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Institution"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/institutions/{id}:
    get:
      operationId: getInstitution
      summary: getInstitution
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Institution"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/institutions/{id}/verification-requests:
    post:
      operationId: requestVerification
      summary: requestVerification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/VerificationResult"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      x-idempotent: true
    get:
      operationId: listVerifications
      summary: listVerifications
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Verification"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/verification-requests/{id}/decisions:
    post:
      operationId: decideVerification
      summary: decideVerification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/VerificationResult"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationDecision"
      x-idempotent: true
  /v1/projects:
    get:
      operationId: listProjects
      summary: listProjects
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          required: false
        - in: query
          name: cursor
          schema:
            type: string
            maxLength: 512
          required: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-paginated: true
    post:
      operationId: createProject
      summary: createProject
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      x-idempotent: true
  /v1/projects/{id}:
    get:
      operationId: getProject
      summary: getProject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateProject
      summary: updateProject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProject"
  /v1/projects/{id}/lifecycle:
    post:
      operationId: transitionProject
      summary: transitionProject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionProject"
      x-idempotent: true
  /v1/projects/{id}/members:
    get:
      operationId: listMembers
      summary: listMembers
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Member"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/projects/{id}/members/{userId}:
    delete:
      operationId: removeMember
      summary: removeMember
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Removed"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VersionCommand"
      x-idempotent: true
  /v1/projects/{id}/ownership-transfer:
    post:
      operationId: transferOwner
      summary: transferOwner
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferOwner"
      x-idempotent: true
...
  /v1/me/invitations:
    get:
      operationId: listInvitations
      summary: listInvitations
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          required: false
        - in: query
          name: cursor
          schema:
            type: string
            maxLength: 512
          required: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invitation"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-paginated: true
  /v1/invitations/{id}:
    get:
      operationId: getInvitation
      summary: getInvitation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Invitation"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/invitations/{id}/accept:
    post:
      operationId: acceptInvitation
      summary: acceptInvitation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/AcceptedInvitation"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-idempotent: true
  /v1/invitations/{id}/decline:
    post:
      operationId: declineInvitation
      summary: declineInvitation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Invitation"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-idempotent: true
  /v1/invitations/{id}/revoke:
    post:
      operationId: revokeInvitation
      summary: revokeInvitation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{8,128}$
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/Invitation"
                  meta:
                    $ref: "#/components/schemas/Meta"
                required:
                  - data
                  - meta
        default:
          description: Structured error; 401 authentication, 403 forbidden, 404 private/missing, 409 conflict, 422 invalid input,
            429 rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-idempotent: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            requestId:
              type: string
          required:
            - code
            - message
            - retryable
            - requestId
      required:
        - error
    Meta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
      required:
        - requestId
...
    ServiceMeta:
      type: object
      additionalProperties: false
      properties:
        apiVersion:
          type: string
        milestones:
          type: array
          items:
            type: string
        features:
          type: object
          additionalProperties: false
          properties:
            projects:
              type: boolean
            invitations:
              type: boolean
            publicProjects:
              type: boolean
            bookings:
              type: boolean
            liveEvents:
              type: boolean
            proposals:
              type: boolean
            progress:
              type: boolean
            resources:
              type: boolean
            availability:
              type: boolean
            consultations:
              type: boolean
            messages:
              type: boolean
            notifications:
              type: boolean
            vouchers:
              type: boolean
            files:
              type: boolean
          required:
            - projects
            - invitations
            - publicProjects
            - bookings
            - liveEvents
        auth:
          type: object
          additionalProperties: false
          properties:
            issuer:
              type: string
            audience:
              type: string
          required:
            - issuer
            - audience
      required:
        - apiVersion
        - milestones
        - features
        - auth
...